// apps/api/src/modules/openai/dto/get-response-dtos/ApplyPatchCallOutputDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsIn, IsOptional, IsString } from 'class-validator';

export class ApplyPatchCallOutputDto {
  /** The unique ID of the apply patch tool call generated by the model. */
  @ApiProperty({ description: `The unique ID of the apply patch tool call generated by the model.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The status of the apply patch tool call output. One of `completed` or `failed`. */
  @ApiProperty({ description: `The status of the apply patch tool call output. One of \`completed\` or \`failed\`.`, enum: ['completed', 'failed'] })
  @IsIn(['completed', 'failed'])
  status!: 'completed' | 'failed';

  /** The type of the item. Always `apply_patch_call_output`. */
  @ApiProperty({ description: `The type of the item. Always \`apply_patch_call_output\`.`, example: 'apply_patch_call_output' })
  @Equals('apply_patch_call_output')
  type!: 'apply_patch_call_output';

  /** The unique ID of the apply patch tool call output. Populated when this item is
returned via API. */
  @ApiProperty({ required: false, description: `The unique ID of the apply patch tool call output. Populated when this item is
returned via API.` })
  @IsOptional()
  id?: null | string;

  /** Optional human-readable log text from the apply patch tool (e.g., patch results
or errors). */
  @ApiProperty({ required: false, description: `Optional human-readable log text from the apply patch tool (e.g., patch results
or errors).` })
  @IsOptional()
  output?: null | string;
}