// apps/api/src/modules/openai/dto/get-response-dtos/ResponseCustomToolCallOutputDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Equals, IsOptional, IsString } from 'class-validator';

import { ResponseInputTextDto } from './ResponseInputTextDto';
import { ResponseInputImageDto } from './ResponseInputImageDto';
import { ResponseInputFileDto } from './ResponseInputFileDto';

export class ResponseCustomToolCallOutputDto {
  /** The call ID, used to map this custom tool call output to a custom tool call. */
  @ApiProperty({ description: `The call ID, used to map this custom tool call output to a custom tool call.`, type: 'string' })
  @IsString()
  call_id!: string;

  /** The output from the custom tool call generated by your code. Can be a string or
an list of output content. */
  @ApiProperty({ description: `The output from the custom tool call generated by your code. Can be a string or
an list of output content.` })
  output!: string | ResponseInputTextDto | ResponseInputImageDto | ResponseInputFileDto[];

  /** The type of the custom tool call output. Always `custom_tool_call_output`. */
  @ApiProperty({ description: `The type of the custom tool call output. Always \`custom_tool_call_output\`.`, example: 'custom_tool_call_output' })
  @Equals('custom_tool_call_output')
  type!: 'custom_tool_call_output';

  /** The unique ID of the custom tool call output in the OpenAI platform. */
  @ApiProperty({ required: false, description: `The unique ID of the custom tool call output in the OpenAI platform.`, type: 'string' })
  @IsOptional()
  @IsString()
  id?: string;
}